Add multipart form handling and helper functions
- Introduced new helper functions for extracting boundaries and parsing multipart form data.
- Updated the file upload middleware to utilize the new multipart parsing logic.
- Enhanced the request body handling to support multipart form data, including file uploads.
- Removed redundant boundary extraction logic from the file upload middleware for cleaner code.
- Added documentation for the new multipart form data handling capabilities.
